move sortRoutes to utils and also sort routes in CanceledTripsContainer
